e2e-gcp-ovn-upgrade to 5h0m (#65209)
4.19+ has more than 100+ tests over previous releases, increasing job
run time to 3h30m to 3h50m, which is on the cusp of timeout -- and often
these jobs do end up timing out if something ends up adding a few
minutes to the run (for example, the current build02 delay to start issue).

This is not permament, we're pursuing other avenues to reduce job
execution time, but this change stops the immediate problem.