dpu-operator: retry checking job status in queue manager (#65930)
Sometimes the curl request seems to fail. As the script intentionally is
silent about what it does, it is not clear why this happens.

Regardless of the problem, it seems a good idea to retry the request a
few times.

Also, after a failure try to re-resolve the IP address. Maybe the
failure happened, because the IP address changed?