sandboxed-containers: fix the pullspecs of the operator (#66424)
* sandboxed-containers: fix the pullspecs of the operator

As we changed the pullspecs from quay.io to registry.redhat.io, we need
the prow jobs to take it into account.

The substitution expects a regular expression that matches the pullspecs
from the CSV file that we create in our builds. It will replace it with
the image pushed to the CI registry during the "build" step of the same
job.

Signed-off-by: Julien Ropé <jrope@redhat.com>

* sandboxed-containers: remove tests for release-1.8 branch

We are not using this branch anymore, so no need to maintain this job.

Signed-off-by: Julien Ropé <jrope@redhat.com>

* sandboxed-containers: use 4.19 as base image

Signed-off-by: Julien Ropé <jrope@redhat.com>

---------

Signed-off-by: Julien Ropé <jrope@redhat.com>