OLMv1: Refresh bumper PR on commit (#68281)
This is to minimize the loss window when a PR is merged while there is
an outstanding bumper PR.

The merge of the bumper PR may cause loss of commits that were merged
between creation and merging of the bumper PR. This attempts to minimize
the window of risk by regenerating the bumper PR when a new commit
occurs.

This should reduce the loss window from (hours/days/weeks) to ~20 mins.

Signed-off-by: Todd Short <todd.short@me.com>